refactor: extract interpreter tests to dedicated file
Move 359 test functions (4,668 lines) from interpreter/mod.rs to
interpreter/tests.rs. Reduces mod.rs from 7,907 to 3,239 lines,
improving navigability of the core interpreter logic.
